requirements.txt
python-telegram-bot>=20.0
python-dotenv
apscheduler
pytz
requests
beautifulsoup4
telethon
cachetools